src/requirements.txt
boto3>=1.28.0
rmscene>=0.7.0
Pillow>=10.0.0
